requirements.txt
flask
qrcode
pillow
pyserial
